requirements.txt
Appium-Python-Client==4.0.0
pytest==8.2.2
python-dotenv==1.0.1
